Rename the file and refector the query statement with using avgMax
