+++ factor-modeling-model/factors/base_factor.py
